Save last search to eventually reuse it

+++ src/herald/iterative_deepening.py
@@ -2,6 +2,7 @@ import multiprocessing
 import time
 from typing import Any, Optional
 
+from . import board
 from .board import Board
 from .configuration import Config
 from .constants import VALUE_MAX
@@ -21,7 +22,7 @@ def to_string(search: Search) -> str:
             if search.time > 0
             else ""
         )
-        + f"pv {' '.join([to_uci(x) for x in search.pv])}"
+        + f"pv {to_uci(search.pv)}"
     )
 
 
@@ -50,29 +51,55 @@ def search_wrapper(
 
 
 def itdep(
+    queue: Any,
     b: Board,
     config: Config,
     movetime: int = 0,
     max_depth: int = 10,
     print_uci: bool = True,
+    last_search: Search | None = None,
 ) -> Optional[Search]:
     # clear transposition table at each new search
     # there seems to be collision/memory issues that I don't have time to handle now
     config.transposition_table.clear()
     config.hash_move_tt.clear()
 
+    if last_search is not None:
+        # try to find a useful subsearch in the last_search
+        for move in last_search.pv:
+            if hash(last_search.board) == hash(b):
+                break
+            # compute next subsearch
+            last_search = Search(
+                board=board.push(last_search.board, move),
+                move=last_search.pv[1],
+                pv=last_search.pv[1:],
+                depth=last_search.depth - 1,
+                nodes=0,
+                score=last_search.score,
+                time=0,
+                stop_search=False,
+            )
+        # clear last_search if it can't be used
+        if hash(last_search.board) != hash(b):
+            last_search = None
+
+    if last_search is not None:
+        start_depth = len(last_search.pv) + 1
+    else:
+        start_depth = 1
+
     if movetime > 0:
         start_time = time.time_ns()
-        last_search: Search | None = None
         max_depth = max(1, min(10, max_depth))
-        for i in range(1, max_depth + 1):
+        for i in range(start_depth, max_depth + 1):
             # we create a queue to be able to stop the search when there's no time left
             # pylint issue: https://github.com/PyCQA/pylint/issues/3488
             # pylint: disable=unsubscriptable-object
-            queue: multiprocessing.Queue[tuple[Search, Config] | None] = multiprocessing.Queue()
+            subqueue: multiprocessing.Queue[tuple[Search, Config] | None] = multiprocessing.Queue()
             process = multiprocessing.Process(
                 target=search_wrapper,
-                args=(queue, b),
+                args=(subqueue, b),
                 kwargs={
                     "depth": i,
                     "config": config,
@@ -89,7 +116,7 @@ def itdep(
             while current_search is None:
                 # we sometime check if we got a move out of the queue
                 try:
-                    ret: tuple[Search, Config] | None = queue.get(True, 0.1)
+                    ret: tuple[Search, Config] | None = subqueue.get(True, 1)
 
                     if ret is not None:
                         (current_search, config) = ret
@@ -104,22 +131,21 @@ def itdep(
                 except:
                     current_search = None
 
-                if isinstance(current_search, Search):
+                if current_search is not None:
                     last_search = current_search
 
-                    # if print_uci:
-                    #     print(to_string(last_search))
-
-                    # bail out if the search tells us to stop
-                    if last_search.stop_search:
-                        process.terminate()
-                        queue.close()
-                        if print_uci:
-                            print(f"bestmove {to_uci(last_search.move)}")
-                        return last_search
+                # bail out if the search tells us to stop
+                if last_search is not None and last_search.stop_search:
+                    process.terminate()
+                    subqueue.close()
+                    queue.put_nowait(last_search)
+                    if print_uci:
+                        print(f"bestmove {to_uci(last_search.move)}")
+                    return last_search
 
                 # bail out if we have a mate
                 if last_search is not None and abs(last_search.score) >= VALUE_MAX:
+                    queue.put_nowait(last_search)
                     if print_uci:
                         print(f"bestmove {to_uci(last_search.move)}")
                     return last_search
@@ -130,12 +156,14 @@ def itdep(
                 # bail out if we have no time anymore
                 if used_time + 1 >= movetime:
                     process.terminate()
-                    queue.close()
+                    subqueue.close()
+                    queue.put_nowait(last_search)
                     if last_search is not None and print_uci:
                         print(f"bestmove {to_uci(last_search.move)}")
                         return last_search
                     return None
 
+        queue.put_nowait(last_search)
         if last_search is not None and print_uci:
             print(f"bestmove {to_uci(last_search.move)}")
             return last_search
@@ -143,7 +171,7 @@ def itdep(
         return None
 
     last_search = None
-    for i in range(1, max_depth + 1):
+    for i in range(start_depth, max_depth + 1):
         ret = search(
             b=b,
             depth=i,
@@ -171,4 +199,5 @@ def itdep(
 
     if last_search is not None and print_uci:
         print(f"bestmove {to_uci(last_search.move)}")
+    queue.put_nowait(last_search)
     return last_search
